Replace react-router TypeScript with JavaScript implementation
