+++ app/routes.js
@@ -1,0 +1,3 @@
+import { index } from "@react-router/dev/routes";
+
+export default [index("routes/home.jsx")];
