+++ package.json
@@ -5,16 +5,18 @@
   "scripts": {
     "build": "react-router build",
     "dev": "react-router dev",
-    "start": "react-router-serve ./build/server/index.js",
-    "typecheck": "react-router typegen && tsc"
+    "start": "react-router-serve ./build/server/index.js"
   },
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@mui/icons-material": "^6.4.8",
     "@mui/material": "^6.4.8",
+    "@mui/x-charts": "^7.28.0",
+    "@mui/x-date-pickers": "^7.28.0",
     "@react-router/node": "^7.4.0",
     "@react-router/serve": "^7.4.0",
+    "dayjs": "^1.11.13",
     "isbot": "^5.1.17",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -23,13 +25,8 @@
   "devDependencies": {
     "@react-router/dev": "^7.4.0",
     "@tailwindcss/vite": "^4.0.0",
-    "@types/node": "^20",
-    "@types/react": "^19.0.1",
-    "@types/react-dom": "^19.0.1",
     "react-router-devtools": "^1.1.0",
     "tailwindcss": "^4.0.0",
-    "typescript": "^5.7.2",
-    "vite": "^5.4.11",
-    "vite-tsconfig-paths": "^5.1.4"
+    "vite": "^5.4.11"
   }
 }
