+++ react-router.config.js
@@ -1,0 +1,5 @@
+export default {
+    // Config options...
+    // Server-side render by default, to enable SPA mode set this to `false`
+    ssr: false,
+};
