--- react-router.config.ts
@@ -1,7 +1,0 @@
-import type { Config } from "@react-router/dev/config";
-
-export default {
-  // Config options...
-  // Server-side render by default, to enable SPA mode set this to `false`
-  ssr: true,
-} satisfies Config;
